requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
dj-database-url==2.3.0
psycopg[binary]==3.2.13
celery[redis]==5.4.0
gunicorn==23.0.0
whitenoise==6.9.0
requests==2.32.3
pypdf==5.4.0
